requirements.txt
asgiref==3.7.2
Django==4.2.4
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
